Rework Event and Task classes
