+++ src/main/java/seedu/task/commons/events/ui/TaskPanelSelectionChangedEvent.java
@@ -1,7 +1,7 @@
 package seedu.task.commons.events.ui;
 
 import seedu.task.commons.events.BaseEvent;
-import seedu.task.model.task.ReadOnlyTask;
+import seedu.task.model.item.ReadOnlyTask;
 
 /**
  * Represents a selection change in the Task List Panel
