+++ src/main/java/seedu/task/logic/commands/SelectCommand.java
@@ -1,7 +1,7 @@
 package seedu.task.logic.commands;
 
 import seedu.task.commons.events.ui.JumpToListRequestEvent;
-import seedu.task.model.task.ReadOnlyTask;
+import seedu.task.model.item.ReadOnlyTask;
 import seedu.taskcommons.core.EventsCenter;
 import seedu.taskcommons.core.Messages;
 import seedu.taskcommons.core.UnmodifiableObservableList;
