--- src/main/java/seedu/task/model/event/Description.java
@@ -1,55 +1,0 @@
-package seedu.task.model.event;
-
-
-import seedu.task.commons.exceptions.IllegalValueException;
-
-/**
- * Represents an Event's description in the task book.
- * Can be used for recording more details about the event.
- * Guarantees: immutable; is valid as declared in {@link #isValidDescription(String)}
- */
-public class Description {
-    
-    public static final String MESSAGE_DESCRIPTION_CONSTRAINTS = "Event descriptions should be spaces or alphanumeric characters";
-    public static final String DESCRIPTION_VALIDATION_REGEX = "[\\p{Alnum} ]+";
-
-    public final String value;
-
-    /**
-     * Validates given description.
-     *
-     * @throws IllegalValueException if given description string is invalid.
-     */
-    public Description(String description) throws IllegalValueException {
-        assert description != null;
-        if (!isValidDescription(description)) {
-            throw new IllegalValueException(MESSAGE_DESCRIPTION_CONSTRAINTS);
-        }
-        this.value = description;
-    }
-
-    /**
-     * Returns true if a given string is a valid description.
-     */
-    public static boolean isValidDescription(String test) {
-        return test.matches(DESCRIPTION_VALIDATION_REGEX);
-    }
-
-    @Override
-    public String toString() {
-        return value;
-    }
-
-    @Override
-    public boolean equals(Object other) {
-        return other == this // short circuit if same object
-                || (other instanceof Description // instanceof handles nulls
-                && this.value.equals(((Description) other).value)); // state check
-    }
-
-    @Override
-    public int hashCode() {
-        return value.hashCode();
-    }
-
-}
