+++ src/main/java/seedu/task/model/item/UniqueEventList.java
@@ -1,4 +1,4 @@
-package seedu.task.model.task;
+package seedu.task.model.item;
 
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
@@ -12,16 +12,16 @@ import java.util.*;
  *
  * Supports a minimal set of list operations.
  *
- * @see Task#equals(Object)
+ * @see Event#equals(Object)
  * @see CollectionUtil#elementsAreUnique(Collection)
  */
-public class UniqueTaskList implements Iterable<Task> {
+public class UniqueEventList implements Iterable<Event> {
 
     /**
      * Signals that an operation would have violated the 'no duplicates' property of the list.
      */
-    public static class DuplicateTaskException extends DuplicateDataException {
-        protected DuplicateTaskException() {
+    public static class DuplicateEventException extends DuplicateDataException {
+        protected DuplicateEventException() {
             super("Operation would result in duplicate tasks");
         }
     }
@@ -30,65 +30,65 @@ public class UniqueTaskList implements Iterable<Task> {
      * Signals that an operation targeting a specified task in the list would fail because
      * there is no such matching task in the list.
      */
-    public static class TaskNotFoundException extends Exception {}
+    public static class EventNotFoundException extends Exception {}
 
-    private final ObservableList<Task> internalList = FXCollections.observableArrayList();
-
-    /**
-     * Constructs empty TaskList.
-     */
-    public UniqueTaskList() {}
+    private final ObservableList<Event> internalList = FXCollections.observableArrayList();
 
     /**
-     * Returns true if the list contains an equivalent task as the given argument.
+     * Constructs empty EventList.
      */
-    public boolean contains(ReadOnlyTask toCheck) {
+    public UniqueEventList() {}
+
+    /**
+     * Returns true if the list contains an equivalent event as the given argument.
+     */
+    public boolean contains(ReadOnlyEvent toCheck) {
         assert toCheck != null;
         return internalList.contains(toCheck);
     }
 
     /**
-     * Adds a task to the list.
+     * Adds a event to the list.
      *
-     * @throws DuplicateTaskException if the task to add is a duplicate of an existing task in the list.
+     * @throws DuplicateEventException if the event to add is a duplicate of an existing event in the list.
      */
-    public void add(Task toAdd) throws DuplicateTaskException {
+    public void add(Event toAdd) throws DuplicateEventException {
         assert toAdd != null;
         if (contains(toAdd)) {
-            throw new DuplicateTaskException();
+            throw new DuplicateEventException();
         }
         internalList.add(toAdd);
     }
 
     /**
-     * Removes the equivalent task from the list.
+     * Removes the equivalent event from the list.
      *
-     * @throws TaskNotFoundException if no such task could be found in the list.
+     * @throws EventNotFoundException if no such event could be found in the list.
      */
-    public boolean remove(ReadOnlyTask toRemove) throws TaskNotFoundException {
+    public boolean remove(ReadOnlyEvent toRemove) throws EventNotFoundException {
         assert toRemove != null;
         final boolean taskFoundAndDeleted = internalList.remove(toRemove);
         if (!taskFoundAndDeleted) {
-            throw new TaskNotFoundException();
+            throw new EventNotFoundException();
         }
         return taskFoundAndDeleted;
     }
 
-    public ObservableList<Task> getInternalList() {
+    public ObservableList<Event> getInternalList() {
         return internalList;
     }
 
     @Override
-    public Iterator<Task> iterator() {
+    public Iterator<Event> iterator() {
         return internalList.iterator();
     }
 
     @Override
     public boolean equals(Object other) {
         return other == this // short circuit if same object
-                || (other instanceof UniqueTaskList // instanceof handles nulls
+                || (other instanceof UniqueEventList // instanceof handles nulls
                 && this.internalList.equals(
-                ((UniqueTaskList) other).internalList));
+                ((UniqueEventList) other).internalList));
     }
 
     @Override
