+++ src/main/java/seedu/task/model/item/UniqueTaskList.java
@@ -1,4 +1,4 @@
-package seedu.task.model.event;
+package seedu.task.model.item;
 
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
@@ -12,16 +12,16 @@ import java.util.*;
  *
  * Supports a minimal set of list operations.
  *
- * @see Event#equals(Object)
+ * @see Task#equals(Object)
  * @see CollectionUtil#elementsAreUnique(Collection)
  */
-public class UniqueEventList implements Iterable<Event> {
+public class UniqueTaskList implements Iterable<Task> {
 
     /**
      * Signals that an operation would have violated the 'no duplicates' property of the list.
      */
-    public static class DuplicateEventException extends DuplicateDataException {
-        protected DuplicateEventException() {
+    public static class DuplicateTaskException extends DuplicateDataException {
+        protected DuplicateTaskException() {
             super("Operation would result in duplicate tasks");
         }
     }
@@ -30,65 +30,65 @@ public class UniqueEventList implements Iterable<Event> {
      * Signals that an operation targeting a specified task in the list would fail because
      * there is no such matching task in the list.
      */
-    public static class EventNotFoundException extends Exception {}
+    public static class TaskNotFoundException extends Exception {}
 
-    private final ObservableList<Event> internalList = FXCollections.observableArrayList();
-
-    /**
-     * Constructs empty EventList.
-     */
-    public UniqueEventList() {}
+    private final ObservableList<Task> internalList = FXCollections.observableArrayList();
 
     /**
-     * Returns true if the list contains an equivalent event as the given argument.
+     * Constructs empty TaskList.
      */
-    public boolean contains(ReadOnlyEvent toCheck) {
+    public UniqueTaskList() {}
+
+    /**
+     * Returns true if the list contains an equivalent task as the given argument.
+     */
+    public boolean contains(ReadOnlyTask toCheck) {
         assert toCheck != null;
         return internalList.contains(toCheck);
     }
 
     /**
-     * Adds a event to the list.
+     * Adds a task to the list.
      *
-     * @throws DuplicateEventException if the event to add is a duplicate of an existing event in the list.
+     * @throws DuplicateTaskException if the task to add is a duplicate of an existing task in the list.
      */
-    public void add(Event toAdd) throws DuplicateEventException {
+    public void add(Task toAdd) throws DuplicateTaskException {
         assert toAdd != null;
         if (contains(toAdd)) {
-            throw new DuplicateEventException();
+            throw new DuplicateTaskException();
         }
         internalList.add(toAdd);
     }
 
     /**
-     * Removes the equivalent event from the list.
+     * Removes the equivalent task from the list.
      *
-     * @throws EventNotFoundException if no such event could be found in the list.
+     * @throws TaskNotFoundException if no such task could be found in the list.
      */
-    public boolean remove(ReadOnlyEvent toRemove) throws EventNotFoundException {
+    public boolean remove(ReadOnlyTask toRemove) throws TaskNotFoundException {
         assert toRemove != null;
         final boolean taskFoundAndDeleted = internalList.remove(toRemove);
         if (!taskFoundAndDeleted) {
-            throw new EventNotFoundException();
+            throw new TaskNotFoundException();
         }
         return taskFoundAndDeleted;
     }
 
-    public ObservableList<Event> getInternalList() {
+    public ObservableList<Task> getInternalList() {
         return internalList;
     }
 
     @Override
-    public Iterator<Event> iterator() {
+    public Iterator<Task> iterator() {
         return internalList.iterator();
     }
 
     @Override
     public boolean equals(Object other) {
         return other == this // short circuit if same object
-                || (other instanceof UniqueEventList // instanceof handles nulls
+                || (other instanceof UniqueTaskList // instanceof handles nulls
                 && this.internalList.equals(
-                ((UniqueEventList) other).internalList));
+                ((UniqueTaskList) other).internalList));
     }
 
     @Override
