+++ src/test/java/guitests/guihandles/TaskCardHandle.java
@@ -3,7 +3,7 @@ package guitests.guihandles;
 import guitests.GuiRobot;
 import javafx.scene.Node;
 import javafx.stage.Stage;
-import seedu.task.model.task.ReadOnlyTask;
+import seedu.task.model.item.ReadOnlyTask;
 
 /**
  * Provides a handle to a task card in the task list panel.
