+++ src/test/java/guitests/guihandles/TaskListPanelHandle.java
@@ -7,8 +7,8 @@ import javafx.scene.Node;
 import javafx.scene.control.ListView;
 import javafx.stage.Stage;
 import seedu.task.TestApp;
-import seedu.task.model.task.Task;
-import seedu.task.model.task.ReadOnlyTask;
+import seedu.task.model.item.ReadOnlyTask;
+import seedu.task.model.item.Task;
 import seedu.task.testutil.TestUtil;
 
 import java.util.List;
