+++ src/test/java/seedu/task/storage/XmlTaskBookStorageTest.java
@@ -9,8 +9,8 @@ import org.junit.rules.TemporaryFolder;
 import seedu.task.commons.exceptions.DataConversionException;
 import seedu.task.commons.util.FileUtil;
 import seedu.task.model.TaskBook;
+import seedu.task.model.item.Task;
 import seedu.task.model.ReadOnlyTaskBook;
-import seedu.task.model.task.Task;
 import seedu.task.storage.XmlTaskBookStorage;
 import seedu.task.testutil.TypicalTestTasks;
